+++ src/steps/fetch-computers/index.ts
@@ -1,0 +1,70 @@
+import {
+  Entity,
+  IntegrationStep,
+  IntegrationStepExecutionContext,
+  createIntegrationEntity,
+} from '@jupiterone/integration-sdk';
+
+import { createDeepSecurityClient, DeepSecurityComputer } from '../../provider';
+
+const COMPUTER_TYPE = 'trend_micro_computer';
+
+const step: IntegrationStep = {
+  id: 'fetch-computers',
+  name: 'Fetch computers',
+  types: [COMPUTER_TYPE],
+  async executionHandler({
+    instance,
+    jobState,
+  }: IntegrationStepExecutionContext) {
+    const client = createDeepSecurityClient(instance);
+
+    const { computers } = await client.listComputers();
+
+    await jobState.addEntities(computers.map(createComputerEntity));
+  },
+};
+
+export default step;
+
+export function createComputerEntity(computer: DeepSecurityComputer): Entity {
+  return createIntegrationEntity({
+    entityData: {
+      source: computer,
+      assign: {
+        _key: createComputerEntityIdentifier(computer),
+        _type: COMPUTER_TYPE,
+        _class: 'Host',
+
+        // normalize property names to match data model
+        createdOn: computer.created,
+        name: computer.displayName || computer.hostName,
+        hostname: computer.hostName,
+        platform: extractPlatform(computer.platform),
+      },
+    },
+  });
+}
+
+/**
+ * DO NOT change this constant. IDs are not long enough
+ * to generate keys that match the min length
+ * the data model requires
+ */
+const API_KEY_ID_PREFIX = 'trend-micro-computer';
+function createComputerEntityIdentifier(
+  computer: DeepSecurityComputer,
+): string {
+  return `${API_KEY_ID_PREFIX}:${computer.ID}`;
+}
+
+/**
+ * This regular expression is used to extract the platform name
+ * that matches our data model from the string returned by the
+ * computer api
+ */
+const PLATFORM_REGEX = /(darwin|linux|unix|windows|android|ios|embedded)/i;
+function extractPlatform(platform: string): string {
+  const [match] = platform.match(PLATFORM_REGEX);
+  return match ? match.toLowerCase() : 'other';
+}